requirements.txt
colorama==0.4.6
mysql_connector_repackaged==0.3.1
Pillow==9.2.0
Pillow==10.2.0
